requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
appdirs==1.4.4
arabic-reshaper==2.1.3
asyncio==3.4.3
asyncio-extras==1.3.2
configparser==5.0.2
discord.py==1.7.3
mal-api==0.4.0
Pillow==8.3.2
praw==7.4.0
prawcore==2.3.0
redditeasy==3.6.3
requests==2.26.0
mysql-connector-python~=8.0.26